DROTH-3578 Refactor RoadAddress to own file, move method getAddressMValuesForCutAssets inside RoadAddress object
