+++ digiroad2-api-oth/src/main/scala/fi/liikennevirasto/digiroad2/LaneApi.scala
@@ -9,8 +9,8 @@ import fi.liikennevirasto.digiroad2.lane.{LanePartitioner, PieceWiseLane}
 import fi.liikennevirasto.digiroad2.linearasset.RoadLink
 import fi.liikennevirasto.digiroad2.service.{RoadAddressService, RoadLinkService}
 import fi.liikennevirasto.digiroad2.util.LaneUtils.pwLanesTwoDigitLaneCode
-import fi.liikennevirasto.digiroad2.util.RoadAddress.isCarTrafficRoadAddress
-import fi.liikennevirasto.digiroad2.util.{PolygonTools, RoadAddress, Track}
+import fi.liikennevirasto.digiroad2.RoadAddress.isCarTrafficRoadAddress
+import fi.liikennevirasto.digiroad2.util.PolygonTools
 import org.json4s.{DefaultFormats, Formats}
 import org.scalatra.{BadRequest, ScalatraServlet}
 import org.scalatra.json.JacksonJsonSupport
