+++ digiroad2-oracle/src/test/scala/fi/liikennevirasto/digiroad2/util/LaneUtilsSpec.scala
@@ -1,5 +1,6 @@
 package fi.liikennevirasto.digiroad2.util
 
+import fi.liikennevirasto.digiroad2.Track
 import fi.liikennevirasto.digiroad2.asset.SideCode
 import fi.liikennevirasto.digiroad2.lane.{LaneEndPoints, LaneRoadAddressInfo}
 import fi.liikennevirasto.digiroad2.service.RoadAddressForLink
